move config and directory to setup
